requirements.txt
pyTelegramBotAPI
SQLAlchemy
psycopg2-binary
python-dotenv
aiogram
asyncpg
APScheduler
python-decouple